+++ js/designer_components/index.ts
@@ -1,0 +1,5 @@
+export { DesignerMultiSelectDropDown } from "./DesignerMultSelectDropDown.ts";
+export { DesignerQuill } from "./DesignerQuill.ts";
+export { DesignerSlider } from "./DesignerSlider.ts";
+export { DesignerSwitch } from "./DesignerSwitch.ts";
+export { DesignerTabs } from "./DesignerTabs.ts";