+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   https://github.com/anvilistas/anvil-extras/pull/64
 * uuid4 in the browser
   https://github.com/anvilistas/anvil-extras/pull/67
+
+## Changes
+* Improved dynamic designer support for Switch, MultiSelectDropDown, Tabs, Quill and Slider
+  https://github.com/anvilistas/anvil-extras/pull/66
+
 # v1.3.1 31-May-2021
 * Improved slider formatting
   https://github.com/anvilistas/anvil-extras/pull/61